Disable LFS (#6)
Recommended by @tfoote since we risk running out of [LFS bandwidth
quota](https://docs.github.com/en/billing/concepts/product-billing/git-lfs#how-use-of-git-lfs-is-measured)
quickly if the repo is downloaded by several users each month.

There is currently a single 3.5MB model file being tracked in LFS, so it
should be benign to convert it to a regular file. We should reconsider
LFS or similar solutions if the repo gets to 100s of MB in size.

---------

Signed-off-by: Shameek Ganguly <shameek@intrinsic.ai>